Add optional 2nd param to nonint do_serial
This makes it possible to disable the serial console without disabling
the UART altogether:

    raspi-config nonint do_serial 1 0

If the first parameter is zero, the second one is ignored.

The second parameter defaults to 1, which means that if you don't supply
it, raspi-config will behave as before.

As discussed in RPi-Distro/raspi-config#29, even without this parameter
it was possible to achieve the same result, but it required two calls:

    raspi-config nonint do_serial 1
    raspi-config nonint set_config_var enable_uart 1 /boot/config.txt