Add debian/copyright file
The Debian Policy Manual says, that every package must be accompanied by a
verbatim copy of its distribution license(s)[1]. The missing copyright file
produces errors when tools try parse these information to check licenses.

Add a copyright file to the debian directory. It repruduces the content
of the LICENSE file in the root directory.

[1] https://www.debian.org/doc/debian-policy/ch-source.html#copyright-debian-copyright

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>