[init_resize.sh] pass --nofsroot to findmnt
For btrfs filesystems mounted with the -subvol option, findmnt appends
the name of the mounted subvolume to the SOURCE column. For example, my
root mountpoint is a subvolume named @root, and it prints
'/dev/mmcblk0p2[/@root]'

This breaks the parsing of partition names in get_variables(). We can
fix this with the --nofsroot option. From the findmnt manpage:
-v, --nofsroot
           Do not print a [/dir] in the SOURCE column for bind mounts or btrfs
           subvolumes.