[init_resize.sh] fix resizepart for btrfs root
The resizepart step fails if the root filesystem is btrfs because
parted prompts for confirmation. Why doesn't it do this for an ext4
root? I do not know.

We can fix this with the same trick used a few lines up in d6b4e8d for
the extended partiton resize.

see also: https://bugs.launchpad.net/ubuntu/+source/parted/+bug/1270203