raspi-config: Change the NetworkManager state file when NetworkManager is enabled but not active
do_wifi_country() does not execute nmcli when systemd.run=/boot/firstrun.sh in cmdline.txt.
In this case, NetworkManager is not active, regardless of whether it is enabled or not.
If wifi is blocked by NetworkManager, it cannot be unblocked on boot.
Therefore, change the NetworkManager state file for this case.